+++ static/js/filetb.js
@@ -78,11 +78,22 @@ $('#id_factor_adjustment, #id_factor_target').focus(function() {
 
 var currentYear = $('#start-year-select').val();
 $('#start-year-select').change(function(e) {
-  $('#current-year-link').attr('href', '/taxbrain/file/?start_year=' + $(this).val());
+  $('#current-year-link').attr('href', '/taxbrain/file/?start_year=' + $(this).val() + '&data_source=' + $('#data-source-select').val());
   $('#current-year-modal').modal('show');
 });
 
 $('#current-year-modal').on('hide.bs.modal', function (e) {
   $('#start-year-select option').removeAttr("selected");
   $('#start-year-select option[value="' + currentYear + '"]').attr("selected", "selected");
+});
+
+var dataSource = $('#data-source-select').val();
+$('#data-source-select').change(function(e) {
+    $('#data-source-link').attr('href', '/taxbrain/file/?start_year=' + $('#start-year-select').val() + '&data_source=' + $(this).val());
+    $('#data-source-modal').modal('show');
+});
+
+$('#data-choice-modal').on('hide.bs.modal', function (e) {
+  $('#data-source option').removeAttr("selected");
+  $('#data-source option[value="' + dataSource + '"]').attr("selected", "selected");
 });