+++ staticfiles/js/main.js
@@ -1,8 +1,0 @@
-/* not currently in use
-$(document).ready(function(){
-  $( "#accordion" ).accordion({ 
-    active: 2,
-    collapsible: true,
-    heightStyle: "content"
-  });
-*/